Fix deletion of change edits
Deletion of change edits only worked accidentally due to a bug that
was introduced by commit 64f0a5e7d57 [formerly 3bde74c2797]. After this bug was fixed by
commit 9330c6d2b18 [formerly 3d227bbadfa] the deletion of change edits was not working
anymore and 2 tests in ChangeEditIT were failing.

Due to the bug DELETE.edit on a change was turned into GET.edit which
resolved to the ChangeEdits collection and this collection was able to
handle the delete. After the bug is fixed a REST view for DELETE.edit
on the ChangesCollection is needed to handle the delete of change
edits.

Change-Id: I4706f5656d27b13dd11561c33ce177cbcb828450
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: e3a82235bf095a51536f892a5e4d49618c91ff4f