Add hashCode method on LabelVote and PatchLineComment classes
According to FindBugs:

"This class overrides equals(Object), but does not override hashCode(),
and inherits the implementation of hashCode() from java.lang.Object
(which returns the identity hash code, an arbitrary value assigned to
the object by the VM).  Therefore, the class is very likely to violate
the invariant that equal objects must have equal hashcodes."

Add implementations of hashCode.

Change-Id: I2d01ecdc51f672b4c9b9d2d1617b3625830f6d7d

Former-commit-id: dc4a36989ac8a061e8cb67730ad90eafbe3e4f57