Simplify parseBody call when looking at changes to close
Every commit is processed through parseBody(), yet the call is in two
locations within the traversal loop.  Clarify the body always gets
parsed and loaded by hoisting it to the top of the loop.

Simplify the loop variable by switching to a for(;;) style syntax,
which scopes the RevCommit variable to only the loop body.

Change-Id: I608f0e3f162ab52f4c24e156e09fadbf40f25d4b

Former-commit-id: 511cb0d5c0f8651c81dde71087840d2b1c66c593