Cache reused values in autoCloseChanges
The tip commit and branch key are passed into the submodule update
helper. Hold them in varibles for reuse, to reduce redundant code.

It is not necessary to reset the RevWalk. The submodule updater is
only using the revision pool to parse commits. It doesn't perform a
traversal, so the reset is merely wasting CPU cycles here.

Change-Id: Id6545033291539b862291eb88d6e8e85a24cb34c

Former-commit-id: c2535f6aa822db36daefdc9ccb0fe7e70c9eb52d