Use changeRefsById to track existing revisions
Most processing paths build up the same sets of ObjectIds, by scanning
through all refs/changes/ and copying them into a temporary set.

Reuse an existing map that is cached after first use and also supports
the autoCloseChanges code path.

It is unlikely that both refs/for/* and refs/heads/* are pushed at the
same time and need to use the map, so there may not be very much reuse
of the data structure within a single execution of ReceiveCommits.

This commit mostly shrinks the code a bit smaller by removing a very
simple redundant processing path.

Change-Id: If2ffd544c821d184593d729fe4ebef77de1860ee

Former-commit-id: 2ffd2cb83f22726fa7274bea54ae065f64064837