Add throwable cause of REST API error to request
When an exception is thrown during processing of a REST API request,
the exception is converted to a simplified error message. The full
contextual detail of the exception may be useful for troubleshooting.
This patchset modifies exception handling routines in the REST API
to store the exception in the request, where it may be accessed by
a debugger or filter.

Change-Id: I6495d3fb72b755b91b77f05de247b2dd03a8fe2c

Former-commit-id: c948ca1093d8507a986a4529bf83677aaf32fcc3