Include all command arguments in SSH log entry
Previously the SSH log only included the first argument. This
prevented the repository name from being logged when
'git receive-pack' was executed instead of 'git-receive-pack'.

Now the SSH log includes all command arguments in the log ensuring
that the repository name is always logged. This is desirable behavior
for anyone looking to monitor repository access via the SSH log.

Change-Id: Idff950e5480a122a2cb366a443d25aa9e0a8f5c8

Former-commit-id: a56057f97620ae7b958c338149853d605d8d3604