Make handling of web links consistent
The WebLinks class now provides all web links as FluentIterables which
avoids unnecessary copying in the caller. Web links that do not have a
name or URL are now always filtered out in the WebLinks class and
callers don't need to do this anymore. In addition, if there are no web
links, the web_links field is not set in the JSON that is sent to
clients.

Implementations of *WebLink can now also return null to indicate that
no web link should be shown for the resource described by the provided
parameters.

Some of these improvements were already done for file web links by
b62414c0c and 2ae832e7e.

Change-Id: I28b98c25ab0f8c4411a6be845181560a636f5e8d

Former-commit-id: e7f3f0a206cf9cee9a9e8decf1061490367b237f