Rename ChangeBatchIndexer to SiteIndexer
The purpose of this class is to reindex the entire site, and it
contains some behaviors that are really only appropriate for this
scope, such as walking repositories to find lists of changed paths
(rather than thrashing PatchListCache), and a synchronous call
returning a results object.

This is in contrast with ReindexAfterUpdate, which operates only on
open changes in a single branch or project with an asynchronous
interface.

In fact, in each case where these different types of batch indexer can
make different choices, they do, which is why they do not share
implementation. The name "ChangeBatchIndexer", however, was too
generic, and suggested that it might be able to fulfill the role of
ReindexAfterUpdate when it can't.

Change-Id: I049746f92d98c642022b3f6a14efe78afe799519

Former-commit-id: db27fb30648d2790ca304ea51fafe187d40106a4