gerrit-pgm: Align Buck rules with package boundaries
Extract http package in its own rule and include it it gerrit-war. That
means that the new artifact is going to be packaged in WEB-INF/lib
directory leaving only root pgm package artifact in pgm-lib.

Remove shell package and move one class in root pgm package. A resource
file that belongs to this functionality as well, was already in root
resource package, so it wasn't consistent anyway.

This change also extracts DEPS dependency definition with artifacts that
are shared in different rules to reuse them and not to repeat the code.

Change-Id: Ie8cbab9c1333ef34dd240619e214674f8b583d2c
Former-commit-id: 919250fd63b80b5f6b90809ec6c3ae61b93e48ef