Enable warnings for missing @Override annotations
Somehow between turning this on and submitting I743917787 I managed to
turn this back off.

Suppress this warning in an instance where a new interface method was
added in Servlet API 3.0, as the code needs to continue to compile
against earlier versions. Unfortunately Eclipse does not give us a
more specific suppression token than "all", but in this case the
method body is small enough we're probably not missing anything.

Change-Id: I78ecf82f6cd4042417c9f3b1f6a4b473d29182d9

Former-commit-id: 54df81ff93ee833ddad76b0301c0c6a3b754ab7a