Allow to pass run options to daemon with gerrit.sh
Option --slave was the only one supported.

Add support in gerrit.sh to pass run options to daemon, which are read
from container.daemonOpt in gerrit.config. I prefer to add generic
option support as opposed to one option in gerrit.config per option
that the daemon supports.

This commit makes the container.slave option obsolete since it can now
be configured in the run options but kept it for backward compatibility.

Change-Id: Id7b9fafc3df055dc6d9d1ef489e9c87e7ba5e470
Former-commit-id: 6710f0a13926791fba2207ad473e644da8a93387