Add InitStep to fix wrong primary key column order
For 2.9 and 2.10 we cannot perform this fix as a database
migration as we cannot increase the schema version. Therefore, do the
primary key fix in an InitStep for 2.9 and 2.10. In the master the
primary key fix will be done as a normal schema migration and the
InitStep introduced here will be deleted.

Change-Id: I4478aabcd63f907a1acfc8e23982bcc6f1c4346f

Former-commit-id: 12b44cb3904ee982c47a8cfbf3499f0f174c67b6