ChangeScreen2: Make Reply button configurable
There were several complaints on Google Groups and in the issue tracker
about the Reply… button in the New Change Screen. People new to
Gerrit/the New Change Screen didn't know where to click in order to
submit a review.

Add a setting in gerrit.config, where the administrator can customize
the label and tooltip of the Reply… button.

Bug: issue 2541
Bug: issue 2974
Change-Id: Ib9e95aa2f03debc88ea091c4d3fb91c83ad9d1ba

Former-commit-id: e4abe9a93d34ce40d561489f40408d650e083f17