Include merge result commit in change-merged stream-event
When CI gets a change merged event what the CI most likely wants
to build is not the last patch set of the change but the result
on the target branch.

If the merge of a change was not fast forward these are two
separate commits.

With this change the CI can retrieve the merge result from the
newRev of a change merged event.

Change-Id: I1d61ab253f4028a26cfbf0427b33873c294ee29e

Former-commit-id: 9686cf224b2bf7abfe1eebe9b942359cb7518396