Only include account IDs when DETAILED_ACCOUNTS is not set
According to the documentation of AccountInfo, only the _accountId is
supposed to be set in the result unless DETAILED_OPTIONS is passed.
The behavior of AccountLoader was inconsistent, only filling the name
in the default case.

Correct the behavior of AccountLoader. The purpose of having detailed
account info be a separate option is to avoid looking up user info in
the cache (possibly incurring a miss) in the simplest case.

Change-Id: Ibcbe5d9733a1dcd2eeca28763438ce87d5fe976d

Former-commit-id: f74f440635184d3ba2edc2db82e9d74d121b6de1