Fix inconsistent behaviour when adding reviewers
It was possible to add reviewers to a change when they had no visibility
to said change and/or were not active users. This is inconsistent with
what happens when adding groups. When adding groups, the non-active
users and the users with no visibility to the change would not be added.
Make the behavior of adding single reviewers consistent with adding
groups.

Change-Id: I606d43e6ca5ada88f26d3b14dd6f16c2306f74af

Former-commit-id: 4248ea8c8626c75995f116bac043d4c04fd24879