Make bug report URL unset by default
There are still many issues [1, 2, 3 and more] being reported on the
Gerrit issue tracker that are not related to Gerrit at all.  These are
most likely from users who have blindly clicked through from a Gerrit
site that is installed with the defaults.

The default URL makes sense on gerrit-review.googlesource.com because
it's most likely that users do actually want to report a bug on Gerrit
itself.

However for most other cases, many users are going to assume that it
links to the issue tracker for the project(s) that is/are being reviewed.
As an example of this, we see quite a lot of issues being reported from
the cyanogenmod review site [4] and the Android open source project's
review site [5].  Both of those are set up with the default URL.

Instead of defaulting to the Gerrit issue tracker, leave the bug report
URL unset if not explicitly specified in the config.

This change effects all installations, newly initialized or already
existing.

The site configuration for gerrit-review.googlesource.com will need to
be updated to explicitly configure the bug report URL.

[1] http://code.google.com/p/gerrit/issues/detail?id=3037
[2] http://code.google.com/p/gerrit/issues/detail?id=3036
[3] http://code.google.com/p/gerrit/issues/detail?id=3028
[4] http://review.cyanogenmod.org/
[5] https://android-review.googlesource.com/

Change-Id: I5c87d456868c37d9bf4cf42dc2ec93beab3444c4
Former-commit-id: 214ab868845bcfb1e33d6665237451916fec0d25