Detecting Trivial Rebases fails unnecessarily early
When label.Label-Name.copyAllScoresOnTrivialRebase
is enabled, the detection of Trivial Rebases fails
unnecessarily early if useContentMerge is disabled
in the project settings.

According to the code comments, a Rebase is Trivial
if the Old PatchSet, cherry picked ontop of Parent
of the New PS, yields the exact same tree as the
New PS.

There is no reason to fail this detection due to
a path conflict regardless of the circumstances.
A Trivial Rebase is always a Trivial Rebase if the
code delta is unchanged.

This fix makes the Trivial Rebase detection ignore
the useContentMerge settings.

Change-Id: I7ef9bda0db3acce0d19bc6888084e57e55cb4b32

Former-commit-id: dd4c0fb16d66bb010ef97d9f12487f350e08595f