Related Changes: Link to ChangeScreen if change is merged or abandoned
For changes in the 'Related Changes' tab that are closed the link was
bringing the user to GitWeb, and not as expected to the ChangeScreen.

As result of this change, merged and abandoned changes are no longer
decorated with the black dot. The decoration will be fixed in a
follow-up change since this requires to transfer the change state to
the client.

Bug: issue 2894
Change-Id: If087a60d08223d9451173d48cca8596f5c13d499
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: d3b6e1d531f0bc6253c6d5ad6205ca1e72ef87f2