Always use side effects of scanning the magic branch
Scanning the magic branch to prepare updates into the Git
batch produces two results, yet the method only returned one of
them. The first result is the newChanges collection. However
as a side-effect the method also updated replaceByChange and
replaceByCommit if a commit matched an existing change using
a Change-Id footer.

Drop the Java method return value and update all of the
collections by side effect.

Change-Id: I6673495e661876f8448611ce70207ef06ed3f31e

Former-commit-id: 729e45f68aac45662a3185cbb04c6eca3c421fe8