Split the ChangeHooks interface in 2
The current ChangeHook interface supports methods for event creators and
listeners.  Since creators and listeners are unlikely to be the same
classes, it does not make sense to expose both sets of methods to each
type of user.  While this change will clean things up API wise, it also
paves the way to split out some event firing logic from the ChangeHooks
class.  By splitting this logic it provides a better path to make it
possible for other classes such as a plugins to fire events.

Change-Id: I0906c86fa6ea32aa519b452134d2caddf489df09

Former-commit-id: bef6c3de3e38722a5493928af17c25ca3020168f