Add postEvent() method that can post any event type
This change provides support for posting and firing events without
having an instance of the original change or branch the event was
created for.

The new postEvent() method is useful for the replication plugin, which
currently does this work itself.  With this change, the replication
plugin can delegate the DB lookups to this new post method.

The new fireEvent() method is useful when re-firing events from another
source such as a plugin which may be deserializing them (along with
supporting the new postEvent() method).

The new isVisibleTo() method is useful to be able to check visibility
when attempting to fire an event to a single user such as when
potentially re-firing old events to them to replay event history (along
with supporting the new fireEvent() method).

Change-Id: Ie4c775d16d8d91aaa3c1e0c70874a10d2fc76fdb

Former-commit-id: 992e5b89a7b09bb16964138f7105b1143520acd7