Fix documentation for merge strategy 'fast-forward-only'
The documentation for merge strategy 'fast-forward-only' asserts that
fast-forward-only produces a strictly linear history. But even with
fast-forward-only it is possible to produce histories not being strictly
linear. E.g. some predecessors of the change being pushed might be
merge commits. The point is that by submitting a change no merge
commit will be created when merge strategy is fast-forward-only.

Change-Id: Ia68e6bac7fb49b1398971e7a23d4277f1ca51b06
Former-commit-id: 7ba3995c3ad938e204f748b58ebe5356209e14ab