Limit number of reductions made by Prolog
Allow site administrators to place an upper bound on how much CPU
time will be spent evaluating rules for a single change. This covers
not just the project rules, but also all filters up to All-Projects.

Reduction counting and not wall-clock time is used to be fair to
rules that are running on a temporarily overloaded system. It also
permits builtin predicates that are expensive like commit_edits/2
to count only once per result made to Prolog, and not charge time
waiting for JGit to populate the PatchListCache.

The default is configured to be 100,000 reductions. This value was
experimentally determined on a Intel Core i7 running at 2.3 GHz and
Java 1.6.0_37. On that system reaching the limit of reductions and
aborting an infinite loop in a user predicate takes about 14 ms.

Any rules taking 14 ms to run will add ~350 ms to the latency of the
All > Open changes screen. Because of this, Administrators already
have sufficient incentives to keep rule running time low.

Change-Id: I2cd5660fc556e6aeeb61606a850ce45c98157c96

Former-commit-id: ed001d7f691e7e4c342141ca06d9efb867db9445