Remove "supported message types" paragraph from stream-events doc
Whenever we add a new stream event we have to mention it in two
places: the "supported message types" list, and the list of events
below.

The former often gets missed and ends up being out of sync with
the latter.

Just remove it. It's enough to document the list of events.

Change-Id: Ib5bab3c508153ce06678b2a452ed9edcf28fe355

Former-commit-id: f3da52f3c0d0caaafcc01c31cdaf52df3e3a2c18