Remove the idea of "release" vs. "developer" schema versions
When prototyping the secondary index during the Boulder hackathon of
June 2013, we originally came up with this idea that certain secondary
index schema versions could not be removed because they were part of a
release, while others could. This idea never really panned out; we
never created even a single developer schema.

Kill that distinction in the code.

Change-Id: I0fd9d4e129a31e444cb2ed6aba2aa75105ba7706

Former-commit-id: b8f05f0c007028ddbcde62c3fe5e5a56aaa80afc